requirements.txt
Flask
google.generativeai
